requirements/dev.txt
django-debug-toolbar
django-tables2